# Skip folded deeper levels when rendering page tree