This change is a performance improvement and should not affect the behavior of the CMS.

Previously the page tree was rendered in full even if pages are folded.
Consider the following tree: even if level 1 was folded, level 2 and 3
would get rendered and sent to the client.

```
- root
 |+ level 1
   |- level 2
     |- level 3
```

This change does two things:

1. load all user-folded pages for the current user in advance, reducing
the number of queries to 1
2. skips all pages that are under a folded page

The result is:

```
- root
 |+ level 1
```
Together with 5a4c0d31e this speeds up the page tree rendering from 20
seconds to ~100ms when only 10 folded pages are shown (for `Alchemy::Page.count == 4000`). The number of queries per page tree is also constant now.

Feedback is welcome! 👂