@pascalj 2017-11-19
Thanks for the gist! Yes, this certainly has an influence. I have no time to dig deeper into this, but it seems Postgres has certain problems with this part of the code or maybe it's my 3.5 branch (although including 5a4c0d3). I tried it with roughly 3k pages (5 levels with 5 sub pages each) and it outperforms my Postgres testsetup. However: both are faster with this patch:

`spec/dummy` with 10 * 100 pages, first level folded:
master: `Completed 200 OK in 1383ms (Views: 1355.2ms | ActiveRecord: 19.0ms)` (~350KB)
this branch: `Completed 200 OK in 244ms (Views: 221.6ms | ActiveRecord: 13.5ms)` (~60KB)

Also notice the difference in size. In the 3.5k pages example this goes down from 600kb to 4KB.

I added a model spec. The test for inheritance doesn't seem to be necessary any more. I copied it from the old implementation, but I'm not brave enough to remove it and added a test to verify it doesn't fail for non-AR user classes.