@creativej 2018-09-07
Great work! What's the timeline for the 4.1 release? 
Any chance for this patch to go into 4.0?
We have a lot of pages in our cms it takes over 5s to load the page tree so getting this improvement would be amazing!